requirements.txt
pip~=22.3.1
wheel~=0.38.4
Jinja2~=3.1.3
numpy~=1.26.4
setuptools~=65.5.1
pytz~=2024.1
cffi~=1.16.0
numba~=0.59.1
packaging~=24.0
librosa~=0.10.1
torch~=2.2.2+cu118
scikit-learn~=1.4.1.post1
pillow~=10.3.0
pyparsing~=3.1.2
matplotlib~=3.8.3
visualkeras~=0.0.2
pytest~=8.1.1
pycore~=17.4.17
opencv-python~=4.9.0.80
seaborn~=0.13.2